+++ appsrc/src/views/vLogin.js
@@ -1,4 +1,4 @@
-// src/views/vLogin.js v0.0.2-1
+// src/views/vLogin.js v0.0.2-2
 /*
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -8,6 +8,8 @@
 var m = require("mithril")
 var vw    = require('../views/vHelper')
 var mUser = require('../models/mUser')
+var mLang = require('../models/mLanguage')
+var S = mLang.useLang('en')
 var Env = {
     ClinicNum: 42
 }
@@ -17,16 +19,10 @@ var vLogin = {
     pass: '',
 }
 
-vLogin.setStatus = id => {
-        return msg => {
-            var node = document.getElementById(id);
-            m.render(node, m('div.mt2.dark-blue.tl',msg))
-        }
-    }
-
 vLogin.setStatus = vw.setById('vLogin_status')
 vLogin.login = () => {
-    vLogin.setStatus('Please wait ...')  // Feedback that request sent
+    // Feedback that request sent
+    vLogin.setStatus(S('pleaseWait')+' ...')
 
     mUser.login(vLogin.user,vLogin.pass)
     .then(() => {
@@ -41,13 +37,13 @@ vLogin.login = () => {
 vLogin.view = () => {
     return m('div.tc',
         m('form.dib.ma3',[
-            m('h1','Clinic #' + Env.ClinicNum + ' Login'),
-            m('label.db.mt2.f6.gray.tl','Username'),
-            m("input[type=text].db.mt1.w-100", {
+            m('h1','Clinic #' + Env.ClinicNum + ' ' + S('login')),
+            m('label.db.mt2.f6.gray.tl',S('username')),
+            m("input[type=text][autofocus].db.mt1.w-100", {
                 oninput: e => {vLogin.user = e.target.value},
                 value: vLogin.user
             }),
-            m('label.db.mt2.f6.gray.tl','Password'),
+            m('label.db.mt2.f6.gray.tl',S('password')),
             m('input[type=password].db.mt1.w-100', {
                 oninput: e => {vLogin.pass = e.target.value},
                 value: vLogin.pass
@@ -58,7 +54,7 @@ vLogin.view = () => {
 
             // Login button
             m('div.mt3',[
-                m('button[type=submit]', {onclick: vLogin.login}, 'Login'),
+                m('button[type=submit]', {onclick: vLogin.login}, S('login')),
             ]),
 
             /* Not implemented